requirements.txt
django
south
django-debug-toolbar
django-extensions
werkzeug

django-social-auth

djangorestframework
markdown
pyyaml
django-filter

requests
